iave

+++ src/pgen/circular_motion_black_hole.cpp
@@ -699,10 +699,53 @@ void MeshBlock::ProblemGenerator(ParameterInput *pin) {
         kl, ku);
   } else {
     peos->PrimitiveToConserved(phydro->w, bb, phydro->u, pcoord, il, iu, jl, ju, kl, ku);
-    bb.DeleteAthenaArray();
+  }
+
+  if (MAGNETIC_FIELDS_ENABLED)
+  ConservedToPrimitive(phydro->u, phydro->w, pfield->b,
+    phydro->w, pfield->bcc, pcoord,il, iu, jl, ju, kl, ku);
+  else 
+  ConservedToPrimitive(phydro->u, phydro->w, pfield->b,
+    phydro->w, bb, pcoord,il, iu, jl, ju, kl, ku);
+
+    // Calculate cell-centered magnetic field
+  if (MAGNETIC_FIELDS_ENABLED) {
+      pfield->CalculateCellCenteredField(pfield->b, pfield->bcc, pcoord, il, iu, jl, ju, kl,
+        ku);
+  } 
+
+  // Initialize conserved values
+  if (MAGNETIC_FIELDS_ENABLED) {
+    peos->PrimitiveToConserved(phydro->w, pfield->bcc, phydro->u, pcoord, il, iu, jl, ju,
+        kl, ku);
+  } else {
+    peos->PrimitiveToConserved(phydro->w, bb, phydro->u, pcoord, il, iu, jl, ju, kl, ku);
+  }
+
+  if (MAGNETIC_FIELDS_ENABLED)
+  ConservedToPrimitive(phydro->u, phydro->w, pfield->b,
+    phydro->w, pfield->bcc, pcoord,il, iu, jl, ju, kl, ku);
+  else 
+  ConservedToPrimitive(phydro->u, phydro->w, pfield->b,
+    phydro->w, bb, pcoord,il, iu, jl, ju, kl, ku);
+
+
+  // Calculate cell-centered magnetic field
+  if (MAGNETIC_FIELDS_ENABLED) {
+      pfield->CalculateCellCenteredField(pfield->b, pfield->bcc, pcoord, il, iu, jl, ju, kl,
+        ku);
+  } 
+
+  // Initialize conserved values
+  if (MAGNETIC_FIELDS_ENABLED) {
+    peos->PrimitiveToConserved(phydro->w, pfield->bcc, phydro->u, pcoord, il, iu, jl, ju,
+        kl, ku);
+  } else {
+    peos->PrimitiveToConserved(phydro->w, bb, phydro->u, pcoord, il, iu, jl, ju, kl, ku);
   }
 
 
+ if (!MAGNETIC_FIELDS_ENABLED) bb.DeleteAthenaArray();
 
   // Free scratch arrays
   g.DeleteAthenaArray();
